(PC-25429)[PRO] refactor: remerge StockFormList into StockEventEdition
